requirements.txt
Flask==2.3.2
tensorflow==2.13.0
keras==2.13.1
Pillow==10.0.0
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.0
